feat: missing disbursement APIs

+++ src/main/java/com/monnify/models/wallet/WalletBalanceResponse.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class WalletBalanceResponse {
-    private Double availableBalance;
-    private Double ledgerBalance;
+    private BigDecimal availableBalance;
+    private BigDecimal ledgerBalance;
 }
